+++ apps/web/lib/apps/conferencing/config.ts
@@ -1,25 +1,6 @@
 import type { App } from "@calcom/types/App";
 
 export const APPS = {
-  huddle01_video: {
-    installed: true,
-    type: "huddle01_video",
-    title: "Huddle01",
-    imageSrc: "integrations/huddle.svg",
-    description: "Video Conferencing",
-    variant: "conferencing",
-    name: "Daily",
-    label: "",
-    slug: "",
-    category: "",
-    logo: "",
-    publisher: "",
-    url: "",
-    verified: true,
-    trending: true,
-    rating: 0,
-    reviews: 0,
-  },
   tandem_video: {
     installed: !!(process.env.TANDEM_CLIENT_ID && process.env.TANDEM_CLIENT_SECRET),
     type: "tandem_video",
