+++ apps/web/lib/random.ts
@@ -1,9 +1,2 @@
-export const randomString = function (length = 12) {
-  let result = "";
-  const characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  const charactersLength = characters.length;
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
-};
+// TODO: Remove this file once everything is imported from `@calcom/lib`
+export * from "@calcom/lib/random";
