+++ apps/web/pages/apps/installed.tsx
@@ -508,7 +508,7 @@ function ConnectOrDisconnectIntegrationButton(props: {
     );
   }
   /** We don't need to "Connect", just show that it's installed */
-  if (props.isGlobal || ["huddle01_video"].includes(props.type)) {
+  if (props.isGlobal) {
     return (
       <div className="truncate px-3 py-2">
         <h3 className="text-sm font-medium text-gray-700">{t("installed")}</h3>
