+++ packages/app-store/huddle01video/lib/index.ts
@@ -1,0 +1,2 @@
+export { default as locationOption } from "./locationOption";
+export { default as VideoApiAdapter } from "./VideoApiAdapter";
