+++ packages/app-store/huddle01video/lib/locationOption.ts
@@ -1,0 +1,9 @@
+import { LocationType } from "@calcom/lib/location";
+
+const locationOption = {
+  value: LocationType.Huddle01,
+  label: "Huddle01 Video",
+  disabled: false,
+};
+
+export default locationOption;
