+++ packages/app-store/index.ts
@@ -1,5 +1,6 @@
 import * as example from "./_example";
 import * as dailyvideo from "./dailyvideo";
+import * as huddle01video from "./huddle01video";
 import * as jitsivideo from "./jitsivideo";
 import * as zoomvideo from "./zoomvideo";
 
@@ -7,6 +8,7 @@ const appStore = {
   example,
   dailyvideo,
   jitsivideo,
+  huddle01video,
   zoomvideo,
 };
